// server/utils/types.ts
export type JsonElement =
  | JsonText
  | JsonButton
  | JsonImage
  | JsonSvg
  | JsonAudio
  | JsonVideo
  | JsonShape
  | JsonEmbed
  | JsonYoutube
  | JsonClipart
  | JsonCountdown
  | JsonSnow
  | JsonRain
  | JsonFog
// ...
export type JsonTextProperties = JsonBaseElementProperties & {
  text?: string
  alignment: JsonTextAlignment
  verticalAlign?: JsonTextVerticalAlignmentType
  contentHeightType: JsonTextContentHeightType
  fontSize: number
  lineHeight: number
  letterSpacing: number
  config: JsonTextSlateConfig // TODO:
  textShadow?: JsonShadow
  initialFontSize?: number
  textDirection?: string | null
  scale?: number
  blendMode?: JsonBlendMode
  outline?: JsonTextOutline
}

export type JsonTextOutline = {
  useOutline?: boolean
  color?: string
  weight?: number
}

export type JsonText = {
  type: 'layer'
  layerType: 'text'
  properties: JsonTextProperties
}
// ...
export type JsonCountdownProperties = JsonBaseElementProperties & {
  countToDate: boolean
  countValue: string | number
  countTimeZone: number
  fontSize: number
  letterSpacing: number
  lineHeight: number
  scale: number
  textShadow: JsonShadow
  alignment: JsonTextAlignment
  config: JsonCountdownConfig[]
}

export type JsonCountdown = {
  type: 'layer'
  layerType: 'countdown'
  properties: JsonCountdownProperties
}

export type JsonRainProperties = JsonBaseElementProperties & {
  size: number
  speed: number
  density: number
  opacity: number
  particleSize: number
  splashEffect: boolean
  splashEffectParticleSize: number
  splashEffectSize: number
}

export type JsonRain = {
  type: 'layer'
  layerType: 'rain'
  properties: JsonRainProperties
}

export type JsonSnowProperties = JsonBaseElementProperties & {
  size: number
  speed: number
  density: number
  opacity: number
  particleSize: number
}

export type JsonSnow = {
  type: 'layer'
  layerType: 'snow'
  properties: JsonSnowProperties
}

export type JsonFogProperties = JsonBaseElementProperties & {
  size: number
  speed: number
  density: number
  opacity: number
}

export type JsonFog = {
  type: 'layer'
  layerType: 'fog'
  properties: JsonFogProperties
}
// ...
export type JsonButtonProperties = JsonBaseElementProperties & {
  buttonLabel: string
  html: string
  backgroundColor?: JsonBackground
  backgroundOverColor?: string
  labelShadow?: JsonShadow
  labelStyle: JsonButtonLabelStyle
  labelOffsetX: number
  labelOffsetY: number
  border?: JsonBorder
  hoverState?: {
    backgroundColor?: JsonBackground
    border?: JsonBorder
    dropShadow?: JsonShadow
    labelShadow?: JsonShadow
    labelStyle?: JsonButtonLabelStyle
  }
}

export type JsonButton = {
  type: 'layer'
  layerType: 'button'
  properties: JsonButtonProperties
}

export type JsonAdjustColor = {
  useAdjustColor: boolean
  brightness: number
  contrast: number
  saturate: number
  hue: number
}

export type JsonCropData = {
  x: number
  y: number
  width: number
  height: number
}

export type JsonImageProperties = JsonBaseElementProperties & {
  url: string
  scaleMode: JsonScaleMode
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  adjustColor: JsonAdjustColor
  contentScale: number
  contentOffsetX: number
  contentOffsetY: number
  originalWidth: number
  originalHeight: number
  hqUrl?: string
  cropData?: JsonCropData
  source?: string
  hash?: string
  licensing?: boolean
  blendMode?: JsonBlendMode
  mediaFolderId?: number
  originalName?: string
  maskImage?: string
}

export type JsonImage = {
  type: 'layer'
  layerType: 'image'
  properties: JsonImageProperties
}

export type JsonLineData = {
  iteration: number
  width: number
  height: number
  left: number
  leftMargin: number
  rightMargin: number
  leftWidth: number
  rightWidth: number
}

export type JsonSvgProperties = JsonBaseElementProperties & {
  url: string
  colorGroups: Record<string, string>
  originalWidth: number
  originalHeight: number
  resourceKey: string
  lineData?: JsonLineData // TODO
}

export type JsonSvg = {
  type: 'layer'
  layerType: 'svg'
  properties: JsonSvgProperties
}

export type JsonBlendMode =
  | 'color'
  | 'color-burn'
  | 'color-dodge'
  | 'darken'
  | 'difference'
  | 'exclusion'
  | 'hard-light'
  | 'hue'
  | 'lighten'
  | 'luminosity'
  | 'multiply'
  | 'normal'
  | 'overlay'
  | 'saturation'
  | 'screen'
  | 'soft-light'

export type JsonShapeProperties = JsonBaseElementProperties & {
  backgroundColor?: JsonBackground
  type: 'circle' | 'rectangle'
  border?: JsonBorder
  cropData?: JsonCropData
  blendMode?: JsonBlendMode
}

export type JsonShape = {
  type: 'layer'
  layerType: 'shape'
  properties: JsonShapeProperties
}

export type JsonYoutubeProperties = JsonBaseElementProperties & {
  youtubeURL: string
  autoplay: boolean
  loop: boolean
  adjustColor?: JsonAdjustColor
  displayYoutubeLogo: boolean
  soundOnMouseOver: boolean
  volume: number
  startAt: boolean
  startAtTime: string
  hideControls: boolean
}

export type JsonYoutube = {
  type: 'layer'
  layerType: 'youtube'
  properties: JsonYoutubeProperties
}

export type JsonUseIframe = 'auto' | 'yes' | 'no'

export type JsonEmbedProperties = JsonBaseElementProperties & {
  backgroundColor?: JsonBackground
  code: string
  useBorder: boolean
  borderColor?: string
  margin: number
  originalWidth?: number
  useIframe: JsonUseIframe
  scrollbar: boolean
  autoResize: boolean
  adjustColor?: JsonAdjustColor
}

export type JsonEmbed = {
  type: 'layer'
  layerType: 'embed'
  properties: JsonEmbedProperties
}

export type JsonAutoplay = 'enabled' | 'disabled'

export type JsonVideoProperties = JsonBaseElementProperties & {
  format: string
  url: string
  scaleMode: JsonScaleMode
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  adjustColor?: JsonAdjustColor
  contentScale: number
  originalWidth: number
  originalHeight: number
  cropData?: JsonCropData
  startTime: number
  endTime: number
  duration: number
  autoplay: JsonAutoplay
  controls: boolean
  loop: boolean
  soundOnHover: boolean
  volume: number
  provider?: JsonStockMediaProvider
  maskImage?: string
  source?: string
}

export type JsonStockMediaProvider =
  | 'unsplash'
  | 'pixabay'
  | 'pexels'
  | 'giphy'
  | 'melodie'
  | 'audioProviders'
  | 'shutterstock'

export type JsonVideo = {
  type: 'layer'
  layerType: 'video'
  properties: JsonVideoProperties
}

export type JsonAudioProperties = JsonBaseElementProperties & {
  format: string
  url: string
  startTime: number
  endTime: number
  duration: number
  volume: number
  soundOnHover: boolean
}

export type JsonAudio = {
  type: 'layer'
  layerType: 'audio'
  properties: JsonAudioProperties
}

export type JsonClipartProperties = JsonBaseElementProperties & {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  svgObject: any
}

export type JsonClipart = {
  type: 'layer'
  layerType: 'clipart'
  properties: JsonClipartProperties
}